Always send directly to user
With the new send, the user is always messages directly by the mite bot. Previously, sends without a name would result in the user getting the message from the slackbot instead of the mite bot.
The new send also removes the need to register with a name (as this isn't needed for sending anymore)