Show the unregister button in case of an error
This seems more helpful than the register button. The user can still unregister first and then register again.
If we only show the register button, and the user can't fix the error by registering again, the app will be stuck trying to send cron trigger messages to the user, but fail.