Make user_id a requirement in the parameters
So that we cannot forget to set it. Not setting it is not what we want most of the time, because that would use the mite information of the user who provided the mite admin key most of the time.
The current user should only be used if the user triggering the action has provided its own api key.